changed users routes
