+++ users_service/__init__.py
@@ -30,6 +30,3 @@ BCRYPT = Bcrypt(APP)
 
 from users_service.views.auth_view import AUTH_BLUEPRINT # pylint: disable=wrong-import-position
 APP.register_blueprint(AUTH_BLUEPRINT)
-
-from users_service.views.users_view import USERS_BLUEPRINT # pylint: disable=wrong-import-position
-APP.register_blueprint(USERS_BLUEPRINT)
